refactor: 重构源码管理配置界面
